Allow fetching over ssh
While we need the github API to enumerate the repositories, there are
advantages to doing the actual repository fetching over ssh.

1. The authentication details are decoupled from the repo. The repo can
be given to any other individual knowing there are no secrets in it.

2. Git is more efficient over ssh than http
(http://git-scm.com/book/ch4-1.html), which could make a big difference
when fetching many repos.

The drawback is that it requires another authentication mechanism to be
setup.