Allow fetching of wikis from github
Wikis for github repos are themselves git repos, but they are not
exposed in the repos API. However, we do know that the wikis are simply
the original repo with ".wiki" appended to the name. If the user has
specified --include-wikis and the repository claims to have a wiki, add
a tweaked wiki repo to the list and fetch it.