Handle failing wiki repo clones
It turns out that the has_wiki attribute of the parent repo is not a
full indicator of the wiki's existence. If the wiki has never been
initialized, then it may not exist and the clone will fail. Catch this
situation by skipping clone errors if we're processing a wiki repo.