Allow the token to be read from a file
Instead of specifying the token on the command line (and leaving it open
for easy snooping), allow the token to be read from a file. If no token
is specified on the command line, then the file at ~/.config/ghtoken
will be read. An alternate file can be specified with the
-t|--token-file option.

Syntax for the token file itself is simply the token string with only
leading and/or trailing whitespace as additional characters.