Rework code and wiki repos into their own classes
This helps pave the way to adding content from the GitHub project that's
not in git format. In that way, each content type can define how it
updates itself instead of piling more utility methods into RepoUpdater.
The WikiRepo tries to reuse the GitRepo methods as much as possible.

The ugliest part is probably in the get_json changes. First, we need to
tell it different object types to use rather than always AttributeDict.
Second, we need to pass through the configuration parameters to the
object, but our object hook doesn't allow for that. We could write a
separate JSON decoder, but this just hacks around it for now.