Refactor repositories to better match GitHub repositories
When the wikis were added, it was treated as another repository from
GitHub. But once that thought is extended to issues, milestones, etc.,
it becomes clear that the idea of a repository would be better to match
how GitHub thinks of a repository: a project with different streams of
content including a code repo.

This change adds the top-level GitHubRepo class, which holds all the
JSON API data and then adds the content repositories. This also means
that the number of repositories again corresponds to the number of
GitHub repositories and not actual git repositories.