Split out API json reading and object construction
We'll reuse these when creating repos of the json data, so split them
out to toplevel functions. This requires that we always pass in the
configuration arguments, so the hack to run the configure method if the
config argument is set has been removed. Instead, we just run the method
directly when creating the GitHubRepos.

Also, the actual of fetching of the json has been split from the object
creation since later we'll just want the json data as is without
creating an object from it.