Add JsonRepo class to handle general repo API content
Unlike the code and wiki that reside in git directories, things like
issues and milestones are just json returned by the API requests. In
these cases, we'll make a simple store that mirrors the API paths. Each
json delivery is saved to a file with the relative API path with .json
suffix.

This class is setup to allow each content type to easily subclass it
with a minimal amount of code. Basically, they'll just have to tell it
what URLs to download.