Don't download pull patches
These aren't served over the API, so they can only be downloaded when
logged in. That's OK for our use case since pulls always come from a
branch in our own repo and the commits will be present in the repo
itself. The information for where to see the pull request is available
in the json.