Put --full-backup content in project.extras directory
Move all the additional content repos under a separate <project>.extras
directory. This keeps the backup directory from getting really cluttered
while retaining the default configuration that code repos go directly
into the toplevel of the backup directory.

Some of the directory creation had to be reworked a bit as it was
relying on being directly under the backup directory.

[endlessm/eos-shell#2038]