Specify directory to clone and run command from cwd
When the code and wiki were being cloned with names following from the
repo (project.git/project.wiki.git), we could rely on git clone to
derive the target directory from the remote URL. However, with the
change to put the wiki in project.extras/wiki.git, this changed.

Instead of letting clone decide the directory to use, specify it
directly and use the current working directory to run the command from.
This will use the target_directory specified by the class.

[endlessm/eos-shell#2038]